requirements.txt
aiohttp==3.5.4
aioredis==1.2.0
PyYAML==5.1


